исправления по замечаниям проверяющего
