+++ package.json
@@ -49,7 +49,7 @@
     "axios": "0.19.2",
     "history": "4.10.1",
     "moment": "2.27.0",
-    "moment-locales-webpack-plugin": "^1.2.0",
+    "moment-locales-webpack-plugin": "1.2.0",
     "react": "16.13.1",
     "react-dom": "16.13.1",
     "react-redux": "7.2.0",
