+++ src/common/consts.js
@@ -27,7 +27,7 @@ export const AppRoute = {
   MYLIST: `/mylist`,
   FILM: `/films`,
   ADD_REVIEW: `/review`,
-  PLAYER: `/player`
+  PLAYER: `/player/`
 };
 
 export const RATING = 5;
