+++ src/components/add-review/add-review.e2e.test.tsx
@@ -41,6 +41,7 @@ describe(`AddReviewComponent`, () => {
               onSubmit={onSubmit}
               onRatingSelect={noop}
               onCommentType={noop}
+              renderMessage={noop}
               isButtonBlocked={false}
               isFormBlocked={false}
               reviewError={false}
