+++ src/hocs/with-input-hadlers/with-input-handlers.tsx
@@ -14,6 +14,7 @@ interface InjectingProps {
     rating: number;
     comment: string;
   }) => void;
+  renderMessage: () => React.ReactNode;
 }
 
 const withInputHandlers = (Component) => {
@@ -29,9 +30,20 @@ const withInputHandlers = (Component) => {
         comment: ``
       };
 
+      this.renderMessage = this.renderMessage.bind(this);
       this.handleSubmit = this.handleSubmit.bind(this);
       this.handleRatingSelect = this.handleRatingSelect.bind(this);
       this.handleCommentType = this.handleCommentType.bind(this);
+    }
+
+    private renderMessage() {
+      return (
+        this.state.comment.length > 0 && this.state.comment.length < Comment.MIN ?
+          <div style={{margin: `0px auto`, fontSize: `15px`, color: `rgba(0, 0, 0, 0.5)`}}>
+            {`Comment should be longer than 50 symbols. ${Comment.MIN - this.state.comment.length} left`}
+          </div> :
+          ``
+      );
     }
 
     private handleRatingSelect(evt) {
@@ -65,6 +77,7 @@ const withInputHandlers = (Component) => {
           onSubmit={this.handleSubmit}
           onRatingSelect={this.handleRatingSelect}
           onCommentType={this.handleCommentType}
+          renderMessage={this.renderMessage}
           isButtonBlocked={this.state.comment.length > Comment.MIN && this.state.rating !== 0 ? false : true}
         />
       );
