+++ src/reducer/data/data.test.js
@@ -1,9 +1,4 @@
-import MockAdapter from "axios-mock-adapter";
-import {createAPI} from "../../api";
-import {reducer, ActionType, Operations} from "./data";
-import filmAdapter from './../../adapter/film';
-
-const api = createAPI(() => {});
+import {reducer, ActionType} from "./data";
 
 const promoFilm = {
   id: 1,
@@ -180,227 +175,5 @@ it(`Reducer should change reveiw status`, () => {
     payload: true,
   })).toEqual({
     reviewSent: true
-  });
-});
-
-describe(`Operation work correctly`, () => {
-  it(`Should make a correct API call to /films/promo`, function () {
-    const apiMock = new MockAdapter(api);
-    const dispatch = jest.fn();
-    const promoFilmLoader = Operations.loadPromoFilm();
-
-    apiMock
-      .onGet(`/films/promo`)
-      .reply(200, [{fake: true}]);
-
-    return promoFilmLoader(dispatch, () => {}, api)
-      .then(() => {
-        expect(dispatch).toHaveBeenCalledTimes(2);
-        expect(dispatch).toHaveBeenNthCalledWith(2, {
-          type: ActionType.LOAD_PROMO_FILM,
-          payload: filmAdapter({fake: true}),
-        });
-      });
-  });
-
-  it(`Should make incorrect API call to /films/promo`, function () {
-    const apiMock = new MockAdapter(api);
-    const dispatch = jest.fn();
-    const promoFilmLoader = Operations.loadPromoFilm();
-
-    apiMock
-      .onGet(`/films/invalid`)
-      .reply(200, [{fake: true}]);
-
-    return promoFilmLoader(dispatch, () => {}, api)
-      .then(() => {
-        expect(dispatch).toHaveBeenCalledTimes(1);
-        expect(dispatch).toHaveBeenNthCalledWith(1, {
-          type: ActionType.SHOW_ERROR,
-          payload: true,
-        });
-      });
-  });
-
-  it(`Should make a correct API call to /films`, function () {
-    const apiMock = new MockAdapter(api);
-    const dispatch = jest.fn();
-    const filmsLoader = Operations.loadFilms();
-
-    apiMock
-      .onGet(`/films`)
-      .reply(200, [{fake: true}]);
-
-    return filmsLoader(dispatch, () => {}, api)
-      .then(() => {
-        expect(dispatch).toHaveBeenCalledTimes(2);
-        expect(dispatch).toHaveBeenNthCalledWith(2, {
-          type: ActionType.LOAD_FILMS,
-          payload: [filmAdapter({fake: true})],
-        });
-      });
-  });
-
-  it(`Should make incorrect API call to /films`, function () {
-    const apiMock = new MockAdapter(api);
-    const dispatch = jest.fn();
-    const filmsLoader = Operations.loadFilms();
-
-    apiMock
-      .onGet(`/invalid`)
-      .reply(200, [{fake: true}]);
-
-    return filmsLoader(dispatch, () => {}, api)
-      .then(() => {
-        expect(dispatch).toHaveBeenCalledTimes(1);
-        expect(dispatch).toHaveBeenNthCalledWith(1, {
-          type: ActionType.SHOW_ERROR,
-          payload: true,
-        });
-      });
-  });
-
-  it(`Should make a correct API call to /comments/film_id`, function () {
-    const apiMock = new MockAdapter(api);
-    const dispatch = jest.fn();
-    const reviewsLoader = Operations.loadReviews(promoFilm);
-
-    apiMock
-      .onGet(`/comments/1`)
-      .reply(200, [{fake: true}]);
-
-    return reviewsLoader(dispatch, () => {}, api)
-      .then(() => {
-        expect(dispatch).toHaveBeenCalledTimes(4);
-        expect(dispatch).toHaveBeenNthCalledWith(4, {
-          type: ActionType.LOAD_REVIEWS,
-          payload: [{fake: true}],
-        });
-      });
-  });
-
-  it(`Should make incorrect API call to /comments/film_id`, function () {
-    const apiMock = new MockAdapter(api);
-    const dispatch = jest.fn();
-    const reviewsLoader = Operations.loadReviews(promoFilm);
-
-    apiMock
-      .onGet(`/comments/1`)
-      .reply(400, [{fake: true}]);
-
-    return reviewsLoader(dispatch, () => {}, api)
-      .then(() => {
-        expect(dispatch).toHaveBeenCalledTimes(2);
-        expect(dispatch).toHaveBeenNthCalledWith(2, {
-          type: ActionType.SHOW_ERROR,
-          payload: true,
-        });
-      });
-  });
-
-  it(`Should make a correct API call to /favorite`, () => {
-    const apiMock = new MockAdapter(api);
-    const dispatch = jest.fn();
-    const favoriteFilmsLoader = Operations.loadFavoriteFilms();
-
-    apiMock
-      .onGet(`/favorite`)
-      .reply(200, [{fake: true}]);
-
-    return favoriteFilmsLoader(dispatch, () => {}, api)
-      .then(() => {
-        expect(dispatch).toHaveBeenCalledTimes(2);
-        expect(dispatch).toHaveBeenNthCalledWith(2, {
-          type: ActionType.LOAD_FAVORITE_FILMS,
-          payload: [filmAdapter({fake: true})],
-        });
-      });
-  });
-
-  it(`Should make incorrect API call to /favorite`, () => {
-    const apiMock = new MockAdapter(api);
-    const dispatch = jest.fn();
-    const favoriteFilmsLoader = Operations.loadFavoriteFilms();
-
-    apiMock
-      .onGet(`/favorite`)
-      .reply(401, [{fake: true}]);
-
-    return favoriteFilmsLoader(dispatch, () => {}, api)
-      .then(() => {
-        expect(dispatch).toHaveBeenCalledTimes(1);
-        expect(dispatch).toHaveBeenNthCalledWith(1, {
-          type: ActionType.SHOW_ERROR,
-          payload: true,
-        });
-      });
-  });
-
-  it(`Should make a correct API call to comments/:id`, () => {
-    const apiMock = new MockAdapter(api);
-    const dispatch = jest.fn();
-    const reviewSender = Operations.sendReview(films[0], reviewData);
-
-    apiMock
-      .onPost(`/comments/2`)
-      .reply(200, [{fake: true}]);
-
-    return reviewSender(dispatch, () => {}, api)
-      .then(() => {
-        expect(dispatch).toHaveBeenCalledTimes(8);
-        expect(dispatch).toHaveBeenNthCalledWith(2, {
-          type: ActionType.SEND_REVIEW,
-          payload: reviewData,
-        });
-      });
-  });
-
-  it(`Should make incorrect API call to comments/:id`, () => {
-    const apiMock = new MockAdapter(api);
-    const dispatch = jest.fn();
-    const reviewSender = Operations.sendReview(films[0], reviewData);
-
-    apiMock
-      .onPost(`/comments/2`)
-      .reply(400, [{fake: true}]);
-
-    return reviewSender(dispatch, () => {}, api)
-      .then(() => {
-        expect(dispatch).toHaveBeenCalledTimes(3);
-        expect(dispatch).toHaveBeenNthCalledWith(3, {
-          type: ActionType.SHOW_REVIEW_ERROR,
-          payload: true,
-        });
-      });
-  });
-
-  it(`Should make a correct API call to favorite/:id/1}`, () => {
-    const apiMock = new MockAdapter(api);
-    const dispatch = jest.fn();
-    const favoriteSender = Operations.changeFavoriteStatus(promoFilm);
-
-    apiMock
-      .onPost(`favorite/1/1`)
-      .reply(200, [{fake: true}]);
-
-    return favoriteSender(dispatch, () => {}, api)
-      .then(() => {
-        expect(dispatch).toHaveBeenCalledTimes(3);
-      });
-  });
-
-  it(`Should make incorrect API call to favorite/:id/1}`, () => {
-    const apiMock = new MockAdapter(api);
-    const dispatch = jest.fn();
-    const favoriteSender = Operations.changeFavoriteStatus(promoFilm);
-
-    apiMock
-      .onPost(`favorite/1/1`)
-      .reply(401, [{fake: true}]);
-
-    return favoriteSender(dispatch, () => {}, api)
-      .then(() => {
-        expect(dispatch).toHaveBeenCalledTimes(1);
-      });
   });
 });
